+++ Source/YaOpt/Helpers/ParallelRegionTraverser.cs
@@ -1,13 +1,19 @@
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.Linq;
+using System.Runtime.CompilerServices;
 using Verse;
+using YaOpt.Patches.Compatibility.VehicleMapFramework;
+using Region = Verse.Region;
 
 namespace YaOpt.Helpers
 {
 	internal static class ParallelRegionTraverser
 	{
 		private static readonly ConcurrentBag<ParallelBFSWorker> pool = new ConcurrentBag<ParallelBFSWorker>();
+
+		internal static bool HasVehicleMapFramework = false;
 
 		static ParallelRegionTraverser()
 		{
@@ -49,7 +55,7 @@ namespace YaOpt.Helpers
 			}
 		}
 
-		private class ParallelBFSWorker
+		internal class ParallelBFSWorker
 		{
 			public void Clear()
 			{
@@ -71,12 +77,22 @@ namespace YaOpt.Helpers
 			{
 			}
 
+			[MethodImpl(MethodImplOptions.AggressiveInlining)]
+			private static bool ValidateRegion(Region from, Region to, HashSet<int> closeSet,
+				RegionEntryPredicate entryCondition, RegionType traversableRegionTypes)
+			{
+				return to != null && !closeSet.Contains(to.id) &&
+					   (to.type & traversableRegionTypes) != RegionType.None &&
+					   (entryCondition == null || entryCondition(from, to));
+			}
+
 			public void BreadthFirstTraverseWork(Region root, RegionEntryPredicate entryCondition, RegionProcessor regionProcessor, int maxRegions, RegionType traversableRegionTypes)
 			{
 				if ((root.type & traversableRegionTypes) == RegionType.None)
 				{
 					return;
 				}
+				var vmf = HasVehicleMapFramework;
 				Clear();
 				numRegionsProcessed = 0;
 				QueueNewOpenRegion(root);
@@ -101,23 +117,91 @@ namespace YaOpt.Helpers
 						FinalizeSearch();
 						return;
 					}
+
+					if (vmf)
+					{
+						VMFPrefix(region, entryCondition, traversableRegionTypes);
+					}
+
 					for (var i = 0; i < region.links.Count; i++)
 					{
 						var regionLink = region.links[i];
 						for (var j = 0; j < 2; j++)
 						{
 							var region2 = regionLink.regions[j];
-							if (region2 != null && !close.Contains(region2.id) &&
-								(region2.type & traversableRegionTypes) != RegionType.None &&
-								(entryCondition == null || entryCondition(region, region2)))
+							if (ValidateRegion(region, region2, close, entryCondition, traversableRegionTypes))
 							{
 								this.QueueNewOpenRegion(region2);
 							}
 						}
 					}
+
+					if (vmf)
+					{
+						VMFPostFix(region, entryCondition, traversableRegionTypes);
+					}
 				}
 				this.FinalizeSearch();
 			}
+
+			#region VehicleMapFramework
+
+			internal void VMFPrefix(Region region, RegionEntryPredicate entryCondition, RegionType traversableRegionTypes)
+			{
+				foreach (var item in region.ListerThings.ThingsInGroup(ThingRequestGroup.Pawn))
+				{
+					if (item is VehicleMapFrameworkCompatibility.VehiclePawnWithMapStub item2)
+					{
+						var region2 = item2.VehicleMap.regionGrid.AllRegions_NoRebuild_InvalidAllowed.FirstOrDefault(r =>
+							ValidateRegion(region, r, close, entryCondition, traversableRegionTypes));
+						if (region2 != null)
+						{
+							QueueNewOpenRegion(region2);
+						}
+					}
+				}
+			}
+
+			internal void VMFPostFix(Region region, RegionEntryPredicate entryCondition, RegionType traversableRegionTypes)
+			{
+				if (!VehicleMapFrameworkCompatibility.IsVehicleMapOfStub(
+						region.Map, out var vehicle))
+				{
+					return;
+				}
+
+				if (vehicle.Spawned)
+				{
+					var regionTo = (vehicle).Position.GetRegion(vehicle.Map, traversableRegionTypes);
+					if (ValidateRegion(region, regionTo, close, entryCondition, traversableRegionTypes))
+					{
+						QueueNewOpenRegion(regionTo);
+						return;
+					}
+				}
+
+				var ziplineDefs = VehicleMapFrameworkCompatibility.ZiplineDefsStub;
+				foreach (var item2 in ziplineDefs.SelectMany(def => region.ListerThings.ThingsOfDef(def)))
+				{
+					var flag = !item2.TryGetComp(out VehicleMapFrameworkCompatibility.CompZiplineStub comp);
+					if (!flag)
+					{
+						var pair = comp.Pair;
+						flag = pair == null || !pair.Spawned;
+					}
+					if (!flag)
+					{
+						var pair2 = comp.Pair;
+						var regionTo = pair2.Position.GetRegion(pair2.Map);
+						if (ValidateRegion(region, regionTo, close, entryCondition, traversableRegionTypes))
+						{
+							QueueNewOpenRegion(regionTo);
+						}
+					}
+				}
+			}
+
+			#endregion
 
 			private Queue<Region> open = new Queue<Region>();